requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
awscli==1.27.59
backcall==0.2.0
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.4
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
discord.py==2.1.0
docutils==0.16
et-xmlfile==1.1.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.14
pyasn1==0.4.8
Pygments==2.7.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
requests-futures==1.0.0
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
stem==1.8.1
torrequest==0.1.0
traitlets==5.0.5
urllib3==1.26.14
wcwidth==0.1.9
yarl==1.8.2
